pyproject.toml
[project]
name = "reddit-watcher"
version = "0.1.0"
description = "Reddit Technical Watcher - A2A agent-based system for monitoring Reddit every 4 hours"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.13",
    "praw>=7.8.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "redis>=6.2.0",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
